moved ghidra into a legacy file to prepare for a pre and post Ghidra 12.0 implementations
